refactor: support .docx file for translation document
